lift proposal helpers to the container level
